+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 All Notable changes to `flipboxdigital\hubspot` will be documented in this file
 
+## Unreleased
+### Added
+- List companies by 'domain'
+
 ## 2.2.0
 ### Added
 - Retrieve a contact by 'utk' token.